build-configs.yaml: add more minimal defconfigs to x86-chromebook
This is a simplification of PR 728 [1] to remove values enabled
by default (via make savedefconfig) and adds a few more configs
to also enable booting Eve (Google Pixelbook).

The x86-chromebook fragment should be as minimal as possible and
remain as close to the mainline x86_64_defconfig as possible but
still allow booting a ChromeOS userspace image and connecting to
the device via for eg. the RTL8152 usb ethernet driver to enable
running some TAST tests.

Some notable features are missing in this fragment like PM, EC,
TPM, HID and posssibly others which could be added in the future
depending on how much divergence from upstream x86_64_defconfig
is acceptable.

Tested on Eve (Adrian) and Coral (Sonny).

[1] https://github.com/kernelci/kernelci-core/pull/728

Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Signed-off-by: Adrian Ratiu <adrian.ratiu@collabora.com>