CHROMEOS test-configs-chromeos.yaml: Disable QEMU
For last 2 releases QEMU is broken for ChromeOS, and it
doesn't look there is easy solution to fix that.
I am commenting temporary relevant entries with hope,
that in new releases it might work or issue might be fixed.
Also commenting is important to not waste lab QEMU resources
and docker pull limits.

Signed-off-by: Denys Fedoryshchenko <denys.f@collabora.com>